Move user-related integration tests to the separate folder
